Bug delete

+++ api/controllers/BenefitController.js
@@ -76,6 +76,9 @@ module.exports = {
   },
   PostbenefitDelete: async function (req, res) {
     const id = req.body.id
+    if (_.isUndefined(id)){
+      return res.badRequest('ID is Undefind.')
+    }
     await Benefit.destroy({
       id: id
     }).exec(function (err) {
