+++ api/controllers/CostdataController.js
@@ -75,7 +75,9 @@ module.exports = {
         }
       },
       PostcostdataDelete: async function (req, res) {
-        const id = req.body.id
+        if (_.isUndefined(id)){
+          return res.badRequest('ID is Undefind.')
+        }
         await Costdata.destroy({
           id: id
         }).exec(function (err) {
