+++ api/controllers/FixcostController.js
@@ -73,7 +73,9 @@ module.exports = {
     }
   },
   PostfixcostDelete: async function (req, res) {
-    const id = req.body.id
+    if (_.isUndefined(id)){
+      return res.badRequest('ID is Undefind.')
+    }
     await Fixcost.destroy({
       id: id
     }).exec(function (err) {
