+++ api/controllers/TeamController.js
@@ -112,7 +112,9 @@ module.exports = {
     }
   },
   PostteamDelete: async function (req, res) {
-    const id = req.body.id
+    if (_.isUndefined(id)){
+      return res.badRequest('ID is Undefind.')
+    }
     await Team.destroy({
       id: id
     }).exec(function (err) {
